[Android] start headless js context on the main thread (#32705)
# Why

On Android our Headless JS context might be initialized from another
thread than the UI thread. Typically it will originate from running
something in response to an action like a remote notification, alarm
etc. which might not be started on the UI thread.

# How

Follow the same pattern as when we do when destroying by calling the
start methods explicitly on the main thread.

# Test Plan

Can be tested with a background task that should be run when a remote
notification is received on Android.

# Checklist

- [X] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>