[Android] Add initialised check when accessing splashScreen property
To avoid crashes if a splash screen is used in a headless (non-activity) context, we need to test if the splash screen variable is set before trying to initialise it.

The use case we're trying to avoid is that we have an app that uses a splash screen and also calls `setOptions()` on load (from JS) in a headless JS context. This is not uncommon when using `expo-task-manager` from push notifications, locations, background-fetch where all of our JS is loaded but we're not creating the splash screen since we don't have an activity.

This PR fixes this by adding a check to see if the lateinit splashScreen property has been initialized.