[cli]: Add expo-router install integration (#32679)
# Why

`expo-router` is unique in the Expo SDK in that it has an explicit
dependency on `@react-navigation/native`. In v4, we require
`@react-navigation/native v7.0.0`. If a project previously had `v6`
installed this may cause `expo-router` to break.

Expo user's normally have the expectation that `npx expo install --fix`
should resolve their projects dependencies, but because this is a unique
situation, the CLI does not detect this issue.

# How

I've added an integration to `@expo/cli` that will call out to
`expo-router/doctor` if the file is available. Expo Router will extend
the dependency check to ensure that its dependancies and
peerDependencies match any packages manually installed.

Additionally, it also checks if `@react-navigation/native` is nested.
This scenario can occur if the user has installed a third party packages
that has a sub dependancy on `@react-navigation/native`. This will cause
two version of `@react-navigation/native` to be installed, possibly
causing issues. For this reason, we compare the resolve path to
`@react-navigation/native` from both the CLI and Expo Router and warn if
there is a difference.

# Test Plan

I've written a unit test and also tested by bundling `@expo/cli`. You
cannot test the sub-dependency logic with `nexpo`, as it will resolve to
the `expo` repro, not your apps repro.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>
Co-authored-by: Brent Vatne <brentvatne@gmail.com>