feat(cli): add `npx expo serve` command (#32602)
# Why

- Create a command to easily run the release server for a project
locally. This is useful for web projects, and React server components.
- In the future, we'll want some `npx expo run` flag which can set the
local server URL as the origin, build the release app, and host the
production server.
- One possibly useful but skewed case would be the ability to host any
`npx expo export` contents, e.g. open a server in Expo Go. The problem
with this is that we should generally keep this script as an extension
of the public `expo/server` that devs will push to production, and the
codesigning aspect of loading in a dev client is not a part of that
(yet?).

<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

# How

- Support `npx expo export` for both output static and dynamic servers.
- Automatically pull from `dist` if no directory is specified.

# Test Plan

- Migrate some E2E tests over to this from the repo scripts.

<!--
Please describe how you tested this change and how a reviewer could
reproduce your test, especially if this PR does not include automated
tests! If possible, please also provide terminal output and/or
screenshots demonstrating your test/reproduction.
-->

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>