feat(metro-config): add scoped module ids for Metro (#32737)
# Why

- Running SSR for different platforms (iOS, web, etc) simultaneously
creates problems when the module IDs are the same value, e.g.
`react/index.js` or `1`. If two platforms update at the same time then
there will be a collision.
- This PR forks around Metro to support passing additional context to
the createModuleId function so the IDs have
`?platform=ios&env=react-server` in them.
- This is also needed for standard web SSR where we have a node and
react-server bundle running in the same process.
- An alternative approach could be to start up different JS threads but
this could make is harder to support different production server
environments such as workerd (cloudflare). We'd also need to ensure the
I/O for each SSR module would need to be serial.
- I tried rewriting the entire module ID system to be non-deterministic
like default Metro, but that doesn't solve the SSR issue since there'd
still be collisions on the numeric IDs.
- split out of https://github.com/expo/expo/pull/32463

# Test Plan

- Existing tests should continue to pass.