[android][background-fetch] change to use execute for tasks Android: The `expo-background-fetch` library starts an alarm when the app is paused to be able to perform a background-task when the app has been backgrounded for a period of time.
If the app was killed after this, the app would've been started in the background in headless mode (without an activity). Instead of executing the defined task in response to being started by the alarm, the old code just scheduled (using the `JobScheduler` API) - resultig in nothing happening untill the app was started by the user and the activity becoming active.

This commit fixes this by doing the right thing - executing the task directly when getting notified by the alarm. This will result in the task being run immediately and if we're in headless mode the app will be loaded by `expo-task-manager` and the `RNHeadlessAppLoader` class.

Here are the details:

**App.tsx**: Added explicitly import of the `BackgroundFetchScreen` to make sure the task is registered also when the screen is not loaded

**BackgroundFetchTaskConsumer.java**: Instead of scheduling the job we're now executing it directly.

**TaskManagerUtilsInterface.java**: Added additional method for executing a task directly

**TaskManagerUtils**: Addedd implementation of executeTask