[android][background-fetch] Fix so that we run tasks correctly when backgrounded or killed (#32849)
# Why

On Android, the `expo-background-fetch` library starts an alarm when the
app is paused to be able to perform a background-task when the app has
been backgrounded for a period of time.

If the app was killed after this, the app would've been started in the
background in headless mode (without an activity). Instead of executing
the defined task in response to being started by the alarm, the old code
just scheduled (using the `JobScheduler` API) - resulting in nothing
happening until the app was started by the user and the activity
becoming active.

# How

This commit fixes this by doing the right thing - executing the task
directly when getting notified by the alarm. This will result in the
task being run immediately and if we're in headless mode the app will be
loaded by `expo-task-manager` and the `RNHeadlessAppLoader` class.

Here are the details:

**App.tsx**: Added explicitly import of the `BackgroundFetchScreen` to
make sure the task is registered also when the screen is not loaded

**BackgroundFetchTaskConsumer.java**: Instead of scheduling the job
we're now executing it directly.

**TaskManagerUtilsInterface.java**: Added additional method for
executing a task directly

**TaskManagerUtils**: Added implementation of executeTask

# Test Plan

Test in BareExpo:
- Run BareExpo (clean install to avoid any async storage data)
- Open API tab
- Open Background Fetch
- Confirm that background fetch has not been run and is not registered
- Click "Register background task"
- Go to home screen (putting app in background will start the background
fetch)
- Test running in the background:
  - Wait approximately 2 minutes 
- Open app and check the BackgroundFetch screen - we should see that it
was called ("Last background fetch....")
- Test running from killed state
  - Kill app using task switcher 
  - Wait approximately 2 minutes 
- Open app and check the BackgroundFetch screen - we should see that it
was called ("Last background fetch....")
 
# Checklist

- [X] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [X] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>
Co-authored-by: Vojtech Novak <vonovak@gmail.com>