[go][ios] Use Go/client icon for launch screen, support dark mode
Why
---
The launch screen shows the Expo logo instead of the Expo Go/dev client icon. It also says "Expo". Remove the text and change the icon to reduce the implication that Expo Go is Expo.

Also the launch screen did not support dark mode.

How
---
Replaced the logo asset with an SVG-based icon. The "any" and "light" icons point to the same file. The tint in light mode is the "slate 1" color we use on our website, which is close to black but not quite and is technically more of a very dark blue. The tint in light mode is plain white.

I also changed the launch screen's background to use the existing backgroundDefault.colorset.

Test Plan
---
Installed the app on my device and tested launching it in light and dark modes.