[router] Fix <Link> using a hash href causing a full page reload on web (#32645)
# Why

Fix `<Link />` using a hash href causing a full page reload on web.

# How

React Navigation does not include support for `#`, this functionality
was added by Expo Router. On web, when an external source causes
navigation (e.g back/forward buttons, anchor tags with hrefs), React
Navigation catches these via the browser navigation events.

For href's with a hash, they actually fire two events, `popstate` and
`hashchange` (in that order). React Navigation only listens to
`popstate`. We could refactor it to also listen to `hashchange`, but
this would be too much work.

The problem logic within in `hashchange` is when it compares the
"history index". The logic within React Navigation doesn't believe an
actionable event has occurred (as the path is the same), so it resets
its state (causing a client side rerender of the page).

I've modified the logic to check that if the hash has changed, React
Navigation should fire an action event and preserve its state.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>