[video][Android] Fix `border` related props (#33075)
# Why

Fixes `border` related props.
A follow up to the https://github.com/expo/expo/pull/33074

# How

The handling of border color and radius was changed in React Native
version 0.76. We are able to patch this in the video package, but I
believe it is better to adopt a standardized method for applying these
properties.

Removed the code responsible for handling border related props. 

# Test Plan

- bare-expo ✅