Enforce Ruby version (#33105)
# Why

@tsapeta and i were digging with the problem for unstable Podfile.lock.
it has been overwritten by different people recently, specifically
hermes-engine and Yoga hash are different.

# How

after digging out, it could be ruby version difference. the
[`Dir.glob`](https://github.com/facebook/react-native/blob/2944752a5de8eba3b240d699baedda83592b9504/packages/react-native/ReactCommon/yoga/Yoga.podspec#L64)
returns sorted result in ruby 3.3 and unsorted result in ruby 2.7.

to this case, it's better to enforce our ruby version. rather than a
`.ruby-version` which is quite strict. i would like to enforce only for
major + minor version. and don't require rbenv, rvm or something. this
pr tries to run `ruby --version` as a sanity check, which would be good
enough.