[docs] update fingerprint example (#33113)
# Why

This updates the fingerprint example in the docs for workflows. This one
should be copy/paste-able. The previous one was not.

# Screenshots

<img width="700" alt="Screenshot 2024-11-20 at 12 02 52 PM"
src="https://github.com/user-attachments/assets/11852315-5a5e-41b6-a3de-b6b8188c888b">


# Test Plan

Make sure the job is as simple as possible to show both local outputs
and global outputs. Also make sure that this workflow works as expected.