[ios][expo-network] fix getNetworkStateAsync failing in SDK52
# Why
On iOS in SDK52 the getNetworkStateAsync fails to return the correct network state.

In SDK-52 we added support for monitoring network changes, resulting in the swift monitor object only being active when someone is listening (through the `useNetworkState` hook or directly).

This causes the `getNetworkStateAsync` function to evaluate wrong if the monitor is not active.

# How
Added a function for retrieving the network state (`getNetworkPathAsync`) was added. This function sets up a temporary monitor and waits for it to finish reporting the network state (with a timeout of 5 seconds).

Closes [SDK52][expo-network] getNetworkStateAsync reports no connection on iOS

Closes #ENG-14290