[sdk-52][ios] fix failing getNetworkStateAsync on sdk-52/iOS (#33137)
# Why
On iOS in SDK52 the getNetworkStateAsync fails to return the correct
network state.

In SDK-52 we added support for monitoring network changes, resulting in
the swift monitor object only being active when someone is listening
(through the `useNetworkState` hook or directly).

This causes the `getNetworkStateAsync` function to evaluate wrong if the
monitor is not active:

<img
src="https://github.com/user-attachments/assets/e84f978e-5ab4-449f-bf64-486bfffbb001"
width=250 />

Observe that the network states are different.

This was reported in #[33012](https://github.com/expo/expo/issues/33012)

# How
Added a function for retrieving the network state
(`getNetworkPathAsync`) was added. This function sets up a temporary
monitor and waits for it to finish reporting the network state (with a
timeout of 5 seconds).

Closes #[ENG-14290](https://linear.app/expo/issue/ENG-14290)

# Test Plan

Use BareExpo and verify that we have a valid network setting are the
same:

<img
src="https://github.com/user-attachments/assets/380701f9-8205-4501-8b0d-dbac8fd4d122"
width=250 />

# Checklist

- [X] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).