[ios][expo-calendar] added reading calendarId when calling createEventInCalendarAsync
The implementation of `createEventInCalendarAsync` did not use the field CalendarId when creating new events. This is a field that should set the calendar field in the iOS dialog for creating/editing an event if provided.

This commit fixes this by reading/using the calendar field. The field has also been added to the test screen in Bare Expo.

Closes #33116