[ios][expo-calendar] added reading calendarId in createEventInCalendarAsync (#33142)
# Why

The implementation of `createEventInCalendarAsync` did not use the field
CalendarId when creating new events. This is a field that should set the
calendar field in the iOS dialog for creating/editing an event if
provided.

Closes #33116

# How

Added reading the calendar field when initializing a new event in Swift
from the JS event.

# Test Plan

Test in BareExpo under API -> Calendar. Select multiple calendars (you
might need to create extra calendars) and press the "Create event using
the OS dialog" button. Ensure that the selected calendar comes up as the
designated calendar in the dialog.

# Checklist

- [X] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>