[android][newarch][task-manager] Added test to avoid destroying host
When the app starts while we're in a task execution context we might have started a headless app which will be taken over by the regular app when it loads.

In the newArch there is a bug in the `RNHeadlessAppLoader`'s `invalidateApp` where it might tear down and destroy while the app is running.

The way this could happen is if a task was triggered by some external service causing the headless app to be setup for the task to run. If the user starts the app at the same time the task will eventually finish and then the reactHost will unfortunately be destroyed by the task manager invalidating the context.

This commit fixes this by adding a similar test to when we run on the old arch to avoid tearing down the reactHost.