[go][iOS] Handle `initialUrl` passed in the `argv` (#33182)
# Why

Handles the `initialUrl` argument when passed as a process argument.

In the Radeon IDE, the application is currently opened via a deep link, which presents a significant drawback: there is no way to access the process's stdout. The IDE team would like to launch the application using `xcrun simctl launch` to access process logs and capture native errors. This pull request adds support for an `initialUrl` parameter passed in the `argv`.

# How

Parsed the `argv` when calculating the `initialUrl`. 

# Test Plan

- launch app from icon ✅ 
- launch app using `xcrun simctl launch` ✅
	```
	xcrun simctl launch <device_id> host.exp.Exponent --initialUrl "exp://192.168.0.235:8081"
	```
- launch app using deep link ✅ 
	```
	xcrun simctl openurl booted exp://192.168.0.235:8081
	```