[docs] point to native upgrade helper for previous versions of Expo Modules manual install guide (#33362)
…es installs

# Why
A number of users transitioning to Expo do so while on earlier versions,
so they isolate their Expo migration from the React Native upgrade. It
might be good in the long run to actually include these previous
instructions, or to even generate the diffs programmatically and include
all supported versions. But for now, let's include a link to what could
help you for previous versions, the native upgrade helper

# How

Added a callout to the native upgrade helper