test(cli): move `max-workers` config to workflow
Its fast enough with 4 different shards on both Windows and Linux. We can optimize the slow tests a bit more, by reusing the caches. And we can clean up the test output to be more verbose with useful information about the commands executed, but lets do that in a follow-up PR.