[autolinking] fix incorrect __dirname in react-native-config
the `react-native.config.(js|ts)` file is loaded via `require-from-string`,
but the `filename` parameter wasn't being passed.
this means that the context regarding the file location was lost,
which results `__dirname` being set to `.` instead of correct location.
this breaks any logic relying on `__dirname` or `__filename`.

so this change adds the filename parameter to fix this issue.
tested in a locally create project.