[screen-orientation][ios] Fix crash with multiple threads accessing member
We've received a Sentry crash (that happens frequently when using expo-screen-orientation) when calling `ScreenOrientation.lockAsync(...)` in app startup.

This never happened in development builds, so I started looking around. The crash report crashed in `Dictionary.isEmpty.getter` in swift - which made me think about concurrent access to the dictionary.

After creating a tight loop that thried to call `ScreenOrientation.lockAsync(...)` 10.000 times after each other (without waiting), I could consistently see the error.

The problem is that while we're reading the dictionary from the main thread, we're writing to it from the module queue's thread.

This commit fixes this by protecting access to the members that aren't thread safe (arrays and dictionaries) using a DispatchQueue that can protect such access.

Closes #33376