[expo-font][ios] fix crash when accessing dictionary from multiple threads
We've received a Sentry crash that was evident for a lot of the client's users.

It wasn't reproducible in development builds, so I started looking around. The crash report crashed in Dictionary.subscript.getter in swift - which made me think about concurrent access to the dictionary.

The fix here protects accessing the `fontFamilyAliases` member by using a DispatchQueue.

Closes #31563