[router] Fix navigation when using browser back/forward (#33524)
<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

Fix: #31747

<!--
How did you build this feature or fix this bug and why?
-->

This issue started after the fork of `createMemoryHistory`
[here](https://github.com/expo/expo/pull/29990/files#diff-8e22bc1cd48d45e4074c0042c548120fde9c1d8af547e622f464579d9f0f3ef9).
The logic that was commented out is necessary to prevent the state from
breaking.

I was thinking we may want to just switch back to importing that
function directly from `@react-navigation/native`, but doing so breaks
e2e tests because it isn't included in react navigation's package
`exports` (due to `Experimental module resolution is enabled`? I'm not
familiar with that yet).

So due to the import issue I just updated the fork of
`createMemoryHistory` to match
[react-navigation](https://github.com/react-navigation/react-navigation/blob/66cc899e2d9264903ed8800980711c619bb3b499/packages/native/src/createMemoryHistory.tsx).

It was originally forked to fix [this
issue](https://github.com/expo/expo/issues/29954), but I confirmed the
e2e tests added for that are still passing. I didn't look too deeply but
I think this is due to the hash being handled correctly
[here](https://github.com/expo/expo/blob/91392d8a29727ffce35eaaeedb7b86f0ae174fb1/packages/expo-router/src/fork/getPathFromState.ts#L300)
now.

The only upstream change since the fork was
[this](https://github.com/react-navigation/react-navigation/pull/11876/files#diff-a26254cf49ae583ad492ff076aa1cd374edb4bb7071618aecfffa38e8e2f1230),
which also seems to relate to not duplicating the hash. The main changes
of that PR were in `useLinking` which we have already updated. I also
added an e2e on our end to cover the issue mentioned in that PR.

While working on this I found a new issue relating to navigating to a
URL with both the hash and query parameters set, but I confirmed it
wasn't introduced by this and I'll open a new ticket for that.

<!--
Please describe how you tested this change and how a reviewer could
reproduce your test, especially if this PR does not include automated
tests! If possible, please also provide terminal output and/or
screenshots demonstrating your test/reproduction.
-->

Added breaking e2e test and then fixed issue

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)

---------

Co-authored-by: Mark Lawlor <mwlawlor@gmail.com>