[video] Add `FlatList` and `FlashList` examples (#33129)
# Why

A lot of people are trying to do TikTok/reels-like apps with
`expo-video` and failing, because they create too many players or create
a new player every time the source changes.
The aim of the screens is to showcase how to implement this in the most
performant way possible. If you have any suggestions please share!

# How

Added two screens - one based on a FlatList and another based on
FlashList. Both work good, FlashList is slightly smoother though. On iOS
it's very good, on Android there is a small hiccup (Edit: source of the
hiccup seems to be a slight corruption of the `ace.mp4` file) when the
player starts playing, I'm still working on the source of it.


# Test Plan

Tested in BareExpo on a Pixel 8 and iPhone 11


https://github.com/user-attachments/assets/848dcb3d-8905-4890-a7a3-571e6b98b355