[expo-modules-core] Ensure `uuid.v4` and `uuid.v5` imports available on old architecture (#33621)
# Why

It seems that `globalThis?.expo?.uuidv4` is undefined on iOS when this
module is loaded. As a result, when trying to call `uuid.v4` later in
the code, it keeps throwing `Native UUID version 4 generator
implementation wasn't found in 'expo-modules-core'` even though
`globalThis?.expo?.uuidv4` at that point is available.

I'm assuming in new architecture `globalThis?.expo?.uuidv4` is set
synchronously, so this would not be an issue. In old architecture
however, it would be best to reference this value when executing `v4` or
`v5` calls themselves.

# How

Use direct reference to `globalThis?.expo?.uuidv4` from inside
`uuidv4()` functions to avoid referencing a stale value.