[dev-menu][dev-launcher] fix broken local development on android (#33714)
# Why

fix broken local development for dev-launcher and dev-menu

# How

there are two issues:

1. since react-native 0.75 (or maybe newer sdk), metro dev server uses
monorepo root to serve bundles. we have to use relative path from
monorepo root for the bundle entry points. relevant changes are like
[#30743](https://github.com/expo/expo/pull/30743/files#diff-3a4872c39f6ba1582f6bb4d0cb259b49720b70cf9925f39143bf08917cee338bR25)
and #31243.
2. for dev-launcher on new arch, we didn't inject devServerHelper in
time and it failed at the
[check](https://github.com/expo/expo/blob/8dd151de7b47b71e158f56f056bc61b0c1c56c50/packages/expo-dev-launcher/android/src/debug/java/expo/modules/devlauncher/helpers/DevLauncherReactUtils.kt#L133).
this pr tries to inject the correct devServerHelper beforehand.