[expo] fix fetch streaming does not complete (#33756)
# Why

fix streaming request does not complete if the request is complete
before streaming started. this is a regression from my #33577

# How

streaming requests rely on event listeners. if the request is complete,
we unregister all event listeners. the caller will never return then.
this pr tries to return the queuedData from startStreaming. for
completed requests, we don't use event listeners anymore.