[docs] Add info on how to enable Array syntax usage in Shared routes (#30964)
# Why

<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

Fixes the issue on how to enable **Array syntax** in an app when using
shared routes.

Refs
[Slack](https://exponent-internal.slack.com/archives/C1QMWQ1P0/p1723151618462689)

# How

<!--
How did you build this feature or fix this bug and why?
-->

Changes:
- On Shared routes page, add info on how to enable array syntax by using
`initialRouteName` in the dynamic layout route file. It continues using
the example defined in the Array syntax section. This will act as a
source of truth on how to enable this setting.
- On Router Settings page, remove the info that explains the context
about using `initialRouteName` for array syntax. This is because the
**Array syntax** section can act as a source of truth in itself and we
should provide the info on how to enable it within that section.

# Test Plan

<!--
Please describe how you tested this change and how a reviewer could
reproduce your test, especially if this PR does not include automated
tests! If possible, please also provide terminal output and/or
screenshots demonstrating your test/reproduction.
-->

Run docs locally and see:
http://localhost:3002/router/advanced/shared-routes/#arrays.

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [x] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).