[sqlite] fix kv-store promises not well transfered when using AsyncSt… (#33847)
# Why

Related to investigations and bug found in
https://github.com/expo/expo/issues/33754.
The promises of the upper function are not transfered to the aliased
function when using AsyncStorage compatible api alias. Await call on
this function are useless and this can trigger database locked when
multiple setItem are trigerred.

# How

Add await and return on the aliased function when needed

(cherry picked from commit d34c1664a962251f10773d793d1c26454be7ce71)