[ios][expo-media-library] Introducing pairedVideoAsset (#33274)
# Why

<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

Currently, when using `expo-media-library`, we don't have access to the
video asset of Live Photos, which would be pretty sweet to have now that
`expo-live-photo` was recently released. This PR attempts to bring this
functionality to the Media Library as well.

I requested this feature and @tsapeta suggested I could try open a PR
for it. Regardless how this goes, I had a lot of fun with this 😄

# How
<!--
How did you build this feature or fix this bug and why?
-->

The implementation adds support for Live Photos by:

- Detecting Live Photos using PHAsset's mediaSubtypes
- Exposing the pairedVideo through a new `pairedVideoAsset`-property in
the `AssetInfo`

I found it reasonable to expose the `pairedVideoAsset` through
`getAssetInfoAsync` - this new property is added to the response of it.
Currently, this property is automatically populated if the image has a
media subtype of `livePhoto`. Would it be better to make this opt-in
instead?

The `PairedVideoAsset` type is designed to look like the
`ImagePickerAsset` type for familiarity.


https://github.com/user-attachments/assets/2812076b-ff7e-49bf-9e41-c960d6657b14

I have not committed the changes to the screen in
`native-components-list` (which you see in the video). That was just a
test I did.

It should also be said that this is my first PR ever touching Swift. I
am not very familiar with it, and not sure about best practices or the
conventions that you may have - so please review it thorough! 😅 Also,
any feedback that you might have is much appreciated - I'd really love
to learn more.

If it's worth continuing with this PR I should also
- update docs
- add a screen to `native-components-list` describing the functionality.

# Test Plan

<!--
Please describe how you tested this change and how a reviewer could
reproduce your test, especially if this PR does not include automated
tests! If possible, please also provide terminal output and/or
screenshots demonstrating your test/reproduction.
-->

I tested this through Bare-Expo on my phone and it seems to work well. 
However, many of the tests for the Media Library seems to fail when I
run them. I will have to look into this.

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
This is required for changes to Expo modules.
-->

- [ ] Documentation is up to date to reflect these changes (eg:
https://docs.expo.dev and README.md).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Tomasz Sapeta <tomasz.sapeta@me.com>