fix(babel): fix server action extraction position (#33933)
# Why

The server actions were being extracted to a single location after the
imports which could lead to nested server actions not being initiated at
the right time. This PR addresses this issue by moving the extraction to
be before the nearest path in the global scope. This is something I've
patched around in the past when using libraries like `ai`.

Fix ENG-14602

# Test Plan

- Added test cases for the syntax.
- Added an extra pass to the tests to ensure the output JS is valid.

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>