[bare-expo] fix search navigation (#33858)
# Why

The navigation via the search screen of the `bare-expo` app is not
working. Seems like a regression from #31992 during upgrading
`@react-navigation`.

# How

Was setting up the local development environment to get familiar with
the internals and noticed this when trying to navigate through the
components and APIs.

# Test Plan

<details>
<summary>Recording before the fix</summary>



https://github.com/user-attachments/assets/1665c222-d8fa-402c-ad10-96a90a6b9a27


</details>

<details>
<summary>Recording after the fix</summary>



https://github.com/user-attachments/assets/0374c16a-495f-44c9-a1fe-f4285502c258


</details>

I've also verified the normal navigation still works as expected.