fix: added guard in expo-router's fast-refresh to avoid fast refresh crashing
When fast-refresh is performed, expo-router overrides the recat-native `isLikelyComponentType` method to provide some extra tests to work with the expo-router's modules.

In #33670 we're seeing that a developer is exporting a stylesheet created with the `react-native-unistyles` library that implements a CPP based stylesheet - meaning that the exported object is a `JSI::HostObject`. This causes fast refresh to fail with the error `WeakSet key must be an object`.

It is a bit unclear why the following test returns true for the Cpp-object:

```ts
if ('unstable_settings' in value)
```

When the following expression returns undefined:

```ts
value["unstable_settings"]
```

The solution provided in this commit is to add a try/catch around the code that inserts elements that it evaluates when fast refres tests if a module exports only react components and is used for _layout files within expo-router.

Closes #ENG-14647