[expo-router] added guard in expo-router's fast-refresh to avoid `WeakSet key must be an object` (#34026)
# Why

In expo-router, when fast-refresh is performed, we extend react-native's
fast-refresh system and overrides the `isLikelyComponentType` method to
provide some extra tests to work with the expo-router's _layout files.

In issue #33670 we're seeing that a developer is exporting a stylesheet
created with the `react-native-unistyles` library that implements a CPP
based stylesheet - meaning that the exported object is a
`JSI::HostObject`. This causes fast refresh to fail with the error
`WeakSet key must be an object`.

I've tried to find out why this doesn't work, without any luck for
JSI::HostObjects:

It is a bit unclear why the following returns true for the Cpp-object:

```ts
if ('unstable_settings' in value)
```

When the following expression returns undefined:

```ts
value["unstable_settings"]
```

# How

The solution provided adds a try/catch around the code that inserts
elements that it evaluates when fast refresh tests if a module exports
only react components and is used for _layout files within expo-router.

Closes #ENG-14647

# Test Plan

Run the provided example in the related issue and verify that it no
longer shows the error message when changing font size (to provoke a
fast refresh).

# Checklist

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).