fix: Restrict color value type to avoid passing illegal PlatformColor
With the current parameter types it is possible to pass `PlatformColor` to `expo-navigation-bar`'s `setBackgroundColorAsync` and `setBorderColorAsync`. This will not work since the native side of the code expectes an int (return value of processColor when not passing in a PlatformColor).

This commit fixes this by changing the typing from ColorValue -> string.

Closes ENG-14609