[] [android][expo-navigation-bar] restrict color value type to avoid passing illegal PlatformColor (#34053)
With the current parameter types it is possible to pass `PlatformColor`
to `expo-navigation-bar`'s `setBackgroundColorAsync` and
`setBorderColorAsync`. This will not work since the native side of the
code expectes an int (return value of processColor when not passing in a
PlatformColor).

This commit fixes this by changing the typing from ColorValue -> string.

Closes ENG-14609

Pass a `PlatformColor` to one of the involved methods to see that you
get ts errors.

- [X] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [X] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).

---------

Co-authored-by: Expo Bot <34669131+expo-bot@users.noreply.github.com>
Co-authored-by: Alan Hughes <30924086+alanjhughes@users.noreply.github.com>