Fix expo-modules-autolinking's reactNativePath not outputting a realpath
When Cocoapods resolves the non-realpath against the react-native
installation, the latter is converted into a realpath. It then tries to
construct relative paths from the symlink folder of `react-native/*` to
the realpath folder of `react-native`. This causes invalid references to
be generated into the `Pods` project, which break the build.

Checking that `reactNativePath` is set to the realpath resolves this.