[expo-constants]: fix regneration of app config
There are scripts in both the podspec and gradle file in expo-constants that recreates the app.config file with values from the app.json file. These scripts are not reliably run when the app is being rebuild on neither Android nor iOS.

This commit fixes this by:

iOS: Added `:always_out_of_date => "1"` to the script phase defining the script in the podspec file

Android: Added `outputs.upToDateWhen { false }` to the gradle task that rebuilds the config file.